Fix directories to delete in the PanCan workflow
Review #1717

+++ grails-app/services/de/dkfz/tbi/otp/ngsdata/LsdfFilesService.groovy
@@ -347,6 +347,9 @@ class LsdfFilesService {
     public void deleteFilesRecursive(Realm realm, Collection<File> filesOrDirectories) {
         assert realm: 'realm may not be null'
         assert filesOrDirectories != null: 'filesOrDirectories may not be null'
+        if (filesOrDirectories.empty) {
+            return //nothing to do
+        }
         String cmd = createClusterScriptService.removeDirs(filesOrDirectories, CreateClusterScriptService.RemoveOption.RECURSIVE_FORCE)
         assert executionService.executeCommand(realm, cmd) ==~ /^0\s*$/
         filesOrDirectories.each {
